made 4

+++ src/main/kotlin/models/Circle.kt
@@ -15,7 +15,7 @@ package models
  * @param startCenter центр окружности
  * @param startRadious радиус окружности
  */
-class Circle(private val startCenter: Point, private val startRadious: Int) {
-    val radious: Int = startRadious
+class Circle(private val startCenter: Point, private val startRadious: Double) {
+    val radious: Double = startRadious
     val center: Point = startCenter
 }
